pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "navigate"
version = "1.0.0"
description = "Multi-point route planning and scheduling optimization system"
requires-python = ">=3.9"
dependencies = [
    "pandas>=1.5",
    "openpyxl>=3.0",
    "xlrd>=2.0",
    "pyyaml>=6.0",
    "requests>=2.28",
    "folium>=0.14",
    "python-docx>=0.8",
]

[project.scripts]
navigate = "navigate.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
